+++ Package.swift
@@ -11,7 +11,15 @@ let package = Package(
             name: "Speakin",
             path: "Sources/Speakin",
             exclude: [
-                "Resources/Info.plist"
+                "Resources/Info.plist",
+                "Resources/AppIcon.icns",
+                "Resources/bird_capsule.png",
+                "Resources/bird_capsule@2x.png",
+                "Resources/bird_menubar.png",
+                "Resources/bird_menubar@2x.png",
+                "Resources/bird_menubar.svg",
+                "Resources/bird_icon_32.png",
+                "Resources/bird_icon_32@2x.png",
             ],
             linkerSettings: [
                 .linkedFramework("AppKit"),